+++ simple_manga_downloader/modules/__init__.py
@@ -21,9 +21,9 @@ def set_mangadex_language(lang_code):
     Mangadex.lang_code = lang_code
 
 
-def set_md_at_home(flag):
-    """Sets the MD@Home setting"""
-    Mangadex.md_at_home = flag
+def set_data_saver(flag):
+    """Sets the data saver setting"""
+    Mangadex.data_saver = flag
 
 
 def toggle_check_only():